# Run Bowtie with --reorder