This ensures that the order of records in Bowtie’s output is the same as the
order of reads in the input FASTQ files.  Without this, the order is more or
less random and we cannot rely on it in the tests.

Closes #101