@marcelm 2019-09-23
Since `--reorder` comes with a performance penalty according to the Bowtie documentation, I considered these alternatives:

- Before checking the MD5 sum, sort by both name and position. This should normalize the order to be the same regardless of what the input order was.

- Use `--threads=1` during the tests. When running Bowtie with only one thread, it will always produce correctly ordered output.

However, I find it a desirable property of a pipeline to always produce the same results, so in my opinion `--reorder` should also be enabled in a production version. On the other hand, we may later want to revisit this and benchmark what the slowdown actually is.