@FrickTobias 2019-09-24
(That rule would then probably be easier to include and make mapper independent by doing something like your first suggestion where files are sorted just before md5sum checks)