@FrickTobias 2019-09-24
Either way, we can implement such an option later if we see the run time it significantly increased (or if we change mapper).