@marcelm 2019-09-25
I have now benchmarked `--reorder` vs not using `--reorder` and there is no difference.

Times for mapping 20 million reads, repeated three times:

* Without `--reorder`: 43m 33s, 43m 22s, 43m 27s
* With `--reorder`: 43m 20s, 43m 14s, 43m 13s